extract ContestAdmin and ContestBoard
